Refactored CheckBoxArray:
- CheckBoxArray is now of type widget and has its own layout.
This way, it can be added as a child widget in any view. Deletion and
memory handling is done by qt with this solution.

Previous solution lead to unwanted behaviour and problems when closing
the app, because the generated radio buttons in CheckBoxArray were used
in layouts of independend views. Since qt manages the memory/ deletion
of all widgets that were added as childs of QWidget in the main app,
problems occured when destructing the CheckBoxArray class and freeing
memory manually.